+++ src/MobileUI/Services/DevService.cs
@@ -4,6 +4,13 @@ using SSW.Rewards.Shared.DTOs.Users;
 
 namespace SSW.Rewards.Mobile.Services;
 
+public interface IDevService
+{
+    Task<IEnumerable<NetworkProfileDto>> GetProfilesAsync();
+    
+    Task<DevProfile> GetProfileAsync(string email);
+}
+
 public class DevService : IDevService
 {
     private readonly IStaffService _staffClient;
