--- src/MobileUI/Services/IDevService.cs
@@ -1,10 +1,0 @@
-﻿using SSW.Rewards.Shared.DTOs.Users;
-
-namespace SSW.Rewards.Mobile.Services;
-
-public interface IDevService
-{
-    Task<IEnumerable<NetworkProfileDto>> GetProfilesAsync();
-    
-    Task<DevProfile> GetProfileAsync(string email);
-}
