--- src/MobileUI/Services/ILeaderService.cs
@@ -1,8 +1,0 @@
-﻿using SSW.Rewards.Shared.DTOs.Leaderboard;
-
-namespace SSW.Rewards.Mobile.Services;
-
-public interface ILeaderService
-{
-    Task<IEnumerable<LeaderboardUserDto>> GetLeadersAsync(bool forceRefresh);
-}
