--- src/MobileUI/Services/IPermissionsService.cs
@@ -1,6 +1,0 @@
-namespace SSW.Rewards.Mobile.Services;
-
-public interface IPermissionsService
-{
-    Task<bool> CheckAndRequestPermission<TPermission>() where TPermission : Permissions.BasePermission, new();
-}
