--- src/MobileUI/Services/IScannerService.cs
@@ -1,6 +1,0 @@
-﻿namespace SSW.Rewards.Mobile.Services;
-
-public interface IScannerService
-{
-    Task<ScanResponseViewModel> ValidateQRCodeAsync(string achievementString);
-}
