--- src/MobileUI/Services/ISnackbarService.cs
@@ -1,8 +1,0 @@
-﻿using SSW.Rewards.Mobile.Controls;
-
-namespace SSW.Rewards.Mobile.Services;
-
-public interface ISnackbarService
-{
-    Task ShowSnackbar(SnackbarOptions options);
-}
