+++ src/MobileUI/Services/QuizService.cs
@@ -3,6 +3,23 @@ using IApiQuizService = SSW.Rewards.ApiClient.Services.IQuizService;
 
 namespace SSW.Rewards.Mobile.Services;
 
+public interface IQuizService
+{
+    Task<IEnumerable<QuizDto>> GetQuizzes();
+
+    Task<QuizDetailsDto> GetQuizDetails(int quizID);
+
+    Task<QuizResultDto> SubmitQuiz(QuizSubmissionDto dto);
+
+    Task<BeginQuizDto> BeginQuiz(int quizId);
+
+    Task SubmitAnswer(SubmitQuizAnswerDto dto);
+
+    Task<bool?> CheckQuizCompletion(int submissionId);
+
+    Task<QuizResultDto> GetQuizResults(int submissionId);
+}
+
 public class QuizService : IQuizService
 {
     private readonly IApiQuizService _quizClient;
