+++ src/MobileUI/Services/SnackBarService.cs
@@ -3,6 +3,11 @@ using SSW.Rewards.Mobile.Controls;
 
 namespace SSW.Rewards.Mobile.Services;
 
+public interface ISnackbarService
+{
+    Task ShowSnackbar(SnackbarOptions options);
+}
+
 public class SnackBarService : ISnackbarService
 {
     public async Task ShowSnackbar(SnackbarOptions options)
